Fix to ITFWriter + Test case
The start data in the ITF encode array was being initialised to the incorrect value with memset. Each item of the array should have the initial value 0x00000001, they were being set to 0x01010101, causing an access error to occur.

The code suffix was also getting set incorrectly and preventing code reading due to a typo