Remove Core module component on QRCode
Core component includes 2 files which are actually from qrcode/multi
Since the framework is supposed to be modular, it should be possible to
exclude QRCode module without breaking the Core component.
Also, even though the headers are imported, the qrcore/module classes
are not used anywhere in Core at all. Those are only used in multi format
reader and writer.