use constant instead of sizeof(array)
This commit fixes 2 occurrences of sizeof(array) which have been used to replace the upstream use of the Java Array.length function.
It happens that in Objective-C sizeof returns the number of the allocated cells while in Java length returns the actual number of entries.
The array here has more allocated than used cells (40 vs 27) which means that the subsequent else clauses are never reached and some symbols like the FNC1 character are interpreted incorrectly. 
A constant value (27) has been used as a solution like in the following else statement. The alternative to set the array allocation to 27 seems to be dangerous in regard to subsequent changes by unaware developers.