Adds support for using the Ultra Wide (macro) camera
Adds an option to request using the Ultra Wide camera if available for better macro support.

This adds the tryUseUltraWideCamera flag on ZXCapture to try to use the Ultra Wide if available. If not available, the Wide camera will be used instead.

This flag must be set before you set the camera property, in order for it to properly take effect.